_my_builder_3/requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
matplotlib>=3.8
pillow>=10.0
reportlab>=4.0
